Move widgets into common UUF component This commit introduces following changes, 	- Move widgets into sample store component. 	- Create sample store provider component. 	- Create sample store provider feature
