+++ components/org.wso2.carbon.dashboards.core/src/main/java/org/wso2/carbon/dashboards/core/internal/dao/impl/DashboardMetadataDAOImpl.java
@@ -23,9 +23,9 @@ import org.slf4j.LoggerFactory;
 import org.wso2.carbon.dashboards.core.bean.DashboardMetadata;
 import org.wso2.carbon.dashboards.core.bean.PaginationContext;
 import org.wso2.carbon.dashboards.core.exception.DashboardException;
-import org.wso2.carbon.dashboards.core.internal.dao.DashboardMetadataDAO;
 import org.wso2.carbon.dashboards.core.internal.dao.utils.DAOUtils;
 import org.wso2.carbon.dashboards.core.internal.dao.utils.SQLConstants;
+import org.wso2.carbon.dashboards.core.internal.dao.DashboardMetadataDAO;
 
 import java.io.IOException;
 import java.sql.Connection;
