+++ components/org.wso2.carbon.dashboards.core/src/main/java/org/wso2/carbon/dashboards/core/internal/provider/impl/DashboardMetadataProviderImpl.java
@@ -31,11 +31,11 @@ import org.slf4j.LoggerFactory;
 import org.wso2.carbon.dashboards.core.bean.DashboardMetadata;
 import org.wso2.carbon.dashboards.core.bean.PaginationContext;
 import org.wso2.carbon.dashboards.core.bean.Query;
+import org.wso2.carbon.dashboards.core.internal.dao.utils.DAOUtils;
+import org.wso2.carbon.dashboards.core.provider.DashboardMetadataProvider;
 import org.wso2.carbon.dashboards.core.exception.DashboardException;
 import org.wso2.carbon.dashboards.core.internal.dao.DashboardMetadataDAO;
 import org.wso2.carbon.dashboards.core.internal.dao.impl.DashboardMetadataDAOImpl;
-import org.wso2.carbon.dashboards.core.internal.dao.utils.DAOUtils;
-import org.wso2.carbon.dashboards.core.provider.DashboardMetadataProvider;
 import org.wso2.carbon.datasource.core.api.DataSourceService;
 
 import java.util.Date;
