--- components/org.wso2.carbon.dashboards.designer/src/main/fragments/bar-chart/conf/widgetConf.json
@@ -1,38 +1,0 @@
-{
-  "info": {
-    "id": "org.wso2.carbon.dashboards.designer.bar-chart",
-    "displayName": "Bar Chart",
-    "thumbnail": "widget-thumbnail.jpg"
-  },
-  "providerConfig": {
-    "stream": "",
-    "version": "",
-    "interval": 500,
-    "type": "",
-    "metadata": {
-      "names": [
-      ],
-      "types": [
-      ]
-    }
-  },
-  "chartConfig": {
-    "type": "",
-    "x": "",
-    "charts": [
-      {
-        "y": ""
-      }
-    ],
-    "maxLength": 20,
-    "width": 800,
-    "height": 400
-  },
-  "pubsub": {
-    "isPublisher": false,
-    "isSubscriber": false,
-    "topic": ""
-  },
-  "permission": {
-  }
-}
