--- components/org.wso2.carbon.dashboards.designer/src/main/fragments/user-pref-sample/public/js/user-pref.js
@@ -1,10 +1,0 @@
-(function () {
-	var bgColor = portal.dashboards.widgets['org.wso2.carbon.dashboards.designer.user-pref-sample'].userpref.bgcolor;
-	$('#content-template').css("background-color", bgColor);
-
-	$( "#color-select" ).change(function(e) {
-		var selectedColor = $("#color-select option:selected").attr('value');
-		portal.dashboards.widgets['org.wso2.carbon.dashboards.designer.user-pref-sample'].userpref.bgcolor = selectedColor;
-		portal.dashboards.functions.view.update();
-	});
-}());
