+++ components/org.wso2.carbon.dashboards.portal/pom.xml
@@ -44,6 +44,13 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.dashboards</groupId>
+            <artifactId>org.wso2.carbon.dashboards.sample.store</artifactId>
+            <version>${carbon.dashboards.version}</version>
+            <type>zip</type>
+            <classifier>uuf-component</classifier>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.dashboards</groupId>
             <artifactId>org.wso2.carbon.dashboards.designer</artifactId>
             <version>${carbon.dashboards.version}</version>
             <type>zip</type>
