+++ components/org.wso2.carbon.dashboards.sample.store/README.md
@@ -1,0 +1,3 @@
+#UUF Dashboard Sample Store Component
+This is a sample store which contains several sample widgets in order to try out dashboard component.
+
