+++ components/org.wso2.carbon.dashboards.sample.store/src/main/component.yaml
@@ -1,0 +1,2 @@
+config:
+    appName: "Sample Widget Store"
