+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/bar-chart/bar-chart.js
