+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/bar-chart/widget-thumbnail.jpg
