+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/pie-chart/pie-chart.hbs
