+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/publisher/publisher.js
