+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/publisher/widget-thumbnail.jpg
