+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/subscriber1/subscriber1.js
