+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/subscriber1/widget-thumbnail.jpg
