+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/user-pref-sample/user-pref-sample.js
