+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/user-pref-sample/widget-thumbnail.jpg
