+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/widget-state/widget-state.hbs
@@ -1,0 +1,12 @@
+<form>
+    <div>
+        <select id="chart-type">
+            <option value="pie">Pie Chart</option>
+            <option value="bar">Bar Chart</option>
+        </select>
+        <button id="btn-draw" type="button">Draw...</button>
+    </div>
+</form>
+
+<div id="chart-canvas"></div>
+<script src="{{public "js/widget-state.js"}}" type="text/javascript"></script>
