+++ components/org.wso2.carbon.dashboards.sample.store/src/main/fragments/widget-state/widget-thumbnail.jpg
